feat(chaining): logic map node creation and placement (#5998)
