+++ openaev-api/src/main/java/io/openaev/service/chaining/WorkflowService.java
@@ -450,6 +450,15 @@ public class WorkflowService {
     workflowRepository.deleteById(workflowId);
   }
 
+  /**
+   * Deletes all workflow states associated with workflows of the given simulation.
+   *
+   * @param simulationId the ID of the simulation whose workflow states should be cleared
+   */
+  public void deleteWorkflowStatesBySimulationId(String simulationId) {
+    workflowStateService.deleteAllBySimulationId(simulationId);
+  }
+
   // -- Configuration Update --
 
   /**
