+++ openaev-front/src/admin/components/chaining/logic/chaining_flow/ChainingFlowConfiguration.tsx
@@ -1,0 +1,250 @@
+import { useCallback, useMemo, useState } from 'react';
+
+import { createStep, updateStep } from '../../../../../actions/chaining/chaining-actions';
+import { useFormatter } from '../../../../../components/i18n';
+import type {
+  ConditionCreateInput,
+  InjectInput,
+  ScopeAssetOutput,
+  ThreatArsenalAction,
+} from '../../../../../utils/api-types';
+import { MESSAGING$ } from '../../../../../utils/Environment';
+import AddActionList from '../drawer/AddActionList';
+import AddComponentDrawer from '../drawer/AddComponentDrawer';
+import ConfigureActionDetail from '../drawer/ConfigureActionDetail';
+import type { ActionDetailData, ActionMeta } from '../types';
+
+export type DrawerView = 'closed' | 'choose' | 'action' | 'actionDetail';
+
+interface ChainingFlowConfigurationProps {
+  workflowId: string | undefined;
+  validAssets: ScopeAssetOutput[];
+  drawerView: DrawerView;
+  onDrawerViewChange: (view: DrawerView) => void;
+  editingStep: {
+    stepId: string;
+    meta: ActionMeta;
+  } | null;
+  onEditingStepChange: (step: {
+    stepId: string;
+    meta: ActionMeta;
+  } | null) => void;
+  onStepCreated: () => void;
+}
+
+const ChainingFlowConfiguration = ({
+  workflowId,
+  validAssets,
+  drawerView,
+  onDrawerViewChange,
+  editingStep,
+  onEditingStepChange,
+  onStepCreated,
+}: ChainingFlowConfigurationProps) => {
+  const { t } = useFormatter();
+
+  // -- Action-specific state --
+  const [selectedAction, setSelectedAction] = useState<ThreatArsenalAction | null>(null);
+
+  // Derive editing data from editingStep
+  const { editingStepId, editingInitialData, editingAction } = useMemo(() => {
+    if (!editingStep) return {
+      editingStepId: null,
+      editingInitialData: undefined,
+      editingAction: null,
+    };
+
+    const { stepId, meta } = editingStep;
+
+    const pseudoAction = {
+      injector_contract_id: meta.inject_injector_contract ?? '',
+      action_labels: { en: meta.inject_title },
+      action_attack_patterns_ids: meta.inject_attack_patterns_ids ?? [],
+      action_injector_type: meta.inject_injector,
+    } as unknown as ThreatArsenalAction;
+
+    const initialData: ActionDetailData = {
+      inject_title: meta.inject_title,
+      inject_injector_contract: meta.inject_injector_contract ?? '',
+      inject_injector: meta.inject_injector,
+      inject_assets: meta.inject_assets ?? [],
+      inject_content: {},
+      inject_field_links: {},
+      contract_fields: meta.contract_fields ?? [],
+    };
+
+    if (meta.step_conditions) {
+      const links: Record<string, {
+        outputType: string;
+        localScope: boolean;
+      }> = {};
+      for (const cond of meta.step_conditions) {
+        if (cond.condition_key) {
+          links[cond.condition_key] = {
+            outputType: cond.condition_key_type ?? 'text',
+            localScope: cond.condition_mapping_type === 'LOCAL',
+          };
+        }
+      }
+      initialData.inject_field_links = links;
+    }
+
+    return {
+      editingStepId: stepId,
+      editingInitialData: initialData,
+      editingAction: pseudoAction,
+    };
+  }, [editingStep]);
+
+  const handleCloseAll = useCallback(() => {
+    onDrawerViewChange('closed');
+    setSelectedAction(null);
+    onEditingStepChange(null);
+  }, [onDrawerViewChange, onEditingStepChange]);
+
+  const handleSelectComponent = (type: 'action' | 'event') => {
+    if (type === 'action') {
+      onDrawerViewChange('action');
+    } else {
+      // TODO: handle event creation drawers
+      onDrawerViewChange('closed');
+    }
+  };
+
+  const handleBackToChoose = () => onDrawerViewChange('choose');
+
+  // -- Actions --
+
+  const handleAddActions = async (selectedActions: ThreatArsenalAction[]) => {
+    if (!workflowId || selectedActions.length === 0) return;
+
+    const promises = selectedActions.map((action) => {
+      const title = action.action_labels?.en
+        ?? action.action_labels?.fr
+        ?? 'Untitled action';
+
+      return createStep({
+        step_workflow_id: workflowId,
+        step_action: 'INJECT_EXECUTION' as const,
+        step_data_step: {
+          inject_title: title,
+          inject_injector_contract: action.injector_contract_id,
+          inject_assets: [],
+          inject_content: {},
+          inject_tags: [],
+          inject_all_teams: false,
+          inject_teams: [],
+          inject_asset_groups: [],
+          inject_documents: [],
+          inject_depends_duration: 0,
+          inject_depends_on: [],
+        } as unknown as InjectInput,
+      });
+    });
+
+    try {
+      const results = await Promise.allSettled(promises);
+      const successCount = results.filter(r => r.status === 'fulfilled').length;
+      if (successCount > 0) {
+        MESSAGING$.notifySuccess(
+          t('{count} action(s) added successfully.', { count: String(successCount) }),
+        );
+      }
+    } finally {
+      onDrawerViewChange('closed');
+      onStepCreated();
+    }
+  };
+
+  const handleSelectAction = (action: ThreatArsenalAction) => {
+    setSelectedAction(action);
+    onDrawerViewChange('actionDetail');
+  };
+
+  const handleBackToActionList = () => {
+    setSelectedAction(null);
+    onDrawerViewChange('action');
+  };
+
+  const handleSaveActionDetail = async (data: ActionDetailData) => {
+    if (!workflowId) return;
+
+    const stepConditions: ConditionCreateInput[] = Object.entries(data.inject_field_links).map(([fieldKey, link], i) => {
+      const parts = link.outputType.split('.');
+      return {
+        condition_temporary_id: String(i),
+        condition_type: 'MAPPER' as const,
+        condition_key_type: parts[0] as ConditionCreateInput['condition_key_type'],
+        condition_key_subtype: (parts.length > 1 ? parts.slice(1).join('.') : undefined) as ConditionCreateInput['condition_key_subtype'],
+        condition_key: fieldKey,
+        condition_mapping_type: (link.localScope ? 'LOCAL' : 'GLOBAL') as ConditionCreateInput['condition_mapping_type'],
+      };
+    });
+
+    const stepPayload = {
+      step_workflow_id: workflowId,
+      step_action: 'INJECT_EXECUTION' as const,
+      step_conditions: stepConditions.length > 0 ? stepConditions : undefined,
+      step_data_step: {
+        inject_title: data.inject_title,
+        inject_injector_contract: data.inject_injector_contract,
+        inject_assets: data.inject_assets,
+        inject_content: data.inject_content,
+        inject_tags: [],
+        inject_all_teams: false,
+        inject_teams: [],
+        inject_asset_groups: [],
+        inject_documents: [],
+        inject_depends_duration: 0,
+        inject_depends_on: [],
+      } as unknown as InjectInput,
+    };
+
+    const savePromise = editingStepId
+      ? updateStep(editingStepId, stepPayload)
+      : createStep(stepPayload);
+
+    savePromise.then(() => {
+      const message = editingStepId
+        ? t('Action updated successfully.')
+        : t('Action added successfully.');
+      MESSAGING$.notifySuccess(message);
+      handleCloseAll();
+      onStepCreated();
+    });
+  };
+
+  // -- Events (future drawers will go here) --
+
+  // Resolve the action to show in ConfigureActionDetail
+  const activeAction = editingAction ?? selectedAction;
+
+  return (
+    <>
+      <AddComponentDrawer
+        open={drawerView === 'choose'}
+        onClose={handleCloseAll}
+        onSelect={handleSelectComponent}
+      />
+      <AddActionList
+        open={drawerView === 'action'}
+        onClose={handleCloseAll}
+        onBack={handleBackToChoose}
+        onAddActions={handleAddActions}
+        onSelectAction={handleSelectAction}
+      />
+      <ConfigureActionDetail
+        open={drawerView === 'actionDetail'}
+        action={activeAction}
+        validAssets={validAssets}
+        initialData={editingInitialData}
+        onClose={handleCloseAll}
+        onBack={editingStepId ? handleCloseAll : handleBackToActionList}
+        onBackToRoot={handleCloseAll}
+        onSave={handleSaveActionDetail}
+      />
+    </>
+  );
+};
+
+export default ChainingFlowConfiguration;
